Split final C++ items out of `wasmtime.hh`
Now `wasmtime.hh` is just an "amalgamation" header to include
everything, yay!
