+++ crates/c-api/tests/simple.cc
@@ -67,115 +67,6 @@ TEST(ExternRef, Smoke) {
   a = b;
 }
 
-TEST(Instance, Smoke) {
-  Engine engine;
-  Store store(engine);
-  Memory m = unwrap(Memory::create(store, MemoryType(1)));
-  Global g = unwrap(Global::create(store, GlobalType(ValKind::I32, false), 1));
-  Table t = unwrap(Table::create(store, TableType(ValKind::FuncRef, 1),
-                                 std::optional<Func>()));
-  Func f(store, FuncType({}, {}),
-         [](auto caller, auto params, auto results) -> auto {
-           return std::monostate();
-         });
-
-  Module mod =
-      unwrap(Module::compile(engine, "(module"
-                                     "(import \"\" \"\" (func))"
-                                     "(import \"\" \"\" (global i32))"
-                                     "(import \"\" \"\" (table 1 funcref))"
-                                     "(import \"\" \"\" (memory 1))"
-
-                                     "(func (export \"f\"))"
-                                     "(global (export \"g\") i32 (i32.const 0))"
-                                     "(export \"m\" (memory 0))"
-                                     "(export \"t\" (table 0))"
-                                     ")"));
-  Instance::create(store, mod, {}).err();
-  Instance i = unwrap(Instance::create(store, mod, {f, g, t, m}));
-  EXPECT_FALSE(i.get(store, "not-present"));
-  f = std::get<Func>(*i.get(store, "f"));
-  m = std::get<Memory>(*i.get(store, "m"));
-  t = std::get<Table>(*i.get(store, "t"));
-  g = std::get<Global>(*i.get(store, "g"));
-
-  EXPECT_TRUE(i.get(store, 0));
-  EXPECT_TRUE(i.get(store, 1));
-  EXPECT_TRUE(i.get(store, 2));
-  EXPECT_TRUE(i.get(store, 3));
-  EXPECT_FALSE(i.get(store, 4));
-  auto [name, func] = *i.get(store, 0);
-  EXPECT_EQ(name, "f");
-}
-
-TEST(Linker, Smoke) {
-  Engine engine;
-  Linker linker(engine);
-  Store store(engine);
-  linker.allow_shadowing(false);
-  Global g = unwrap(Global::create(store, GlobalType(ValKind::I32, false), 1));
-  unwrap(linker.define(store, "a", "g", g));
-  unwrap(linker.define_wasi());
-  unwrap(linker.func_new("a", "f", FuncType({}, {}),
-                         [](auto caller, auto params, auto results) -> auto {
-                           return std::monostate();
-                         }));
-  unwrap(linker.func_wrap("a", "f2", []() {}));
-  unwrap(linker.func_wrap("a", "f3", [](Caller arg) {}));
-  unwrap(linker.func_wrap("a", "f4", [](Caller arg, int32_t a) {}));
-  Module mod = unwrap(Module::compile(engine, "(module)"));
-  Instance i = unwrap(Instance::create(store, mod, {}));
-  unwrap(linker.define_instance(store, "x", i));
-  unwrap(linker.instantiate(store, mod));
-  unwrap(linker.module(store, "y", mod));
-  EXPECT_TRUE(linker.get(store, "a", "g"));
-  unwrap(linker.get_default(store, "g"));
-  EXPECT_TRUE(linker.get(store, "a", "f"));
-  EXPECT_TRUE(std::holds_alternative<Func>(*linker.get(store, "a", "f")));
-}
-
-TEST(Linker, CallableMove) {
-  Engine engine;
-  Linker linker(engine);
-  Store store(engine);
-  linker.allow_shadowing(false);
-
-  struct CallableFunc {
-    CallableFunc() = default;
-    CallableFunc(const CallableFunc &) = delete;
-    CallableFunc(CallableFunc &&) = default;
-
-    Result<std::monostate, Trap>
-    operator()(Caller caller, Span<const Val> params, Span<Val> results) {
-      return std::monostate();
-    }
-  };
-
-  CallableFunc cf;
-  unwrap(linker.func_new("a", "f", FuncType({}, {}), std::move(cf)));
-}
-
-TEST(Linker, CallableCopy) {
-  Engine engine;
-  Linker linker(engine);
-  Store store(engine);
-  linker.allow_shadowing(false);
-
-  struct CallableFunc {
-    CallableFunc() = default;
-    CallableFunc(const CallableFunc &) = default;
-    CallableFunc(CallableFunc &&) = default;
-
-    Result<std::monostate, Trap>
-    operator()(Caller caller, Span<const Val> params, Span<Val> results) {
-      return std::monostate();
-    }
-  };
-
-  CallableFunc cf;
-  unwrap(linker.func_new("a", "f", FuncType({}, {}), cf));
-}
-
 TEST(Caller, Smoke) {
   Engine engine;
   Store store(engine);
